ARC: int vs long
- arch specific bitsperlong.h
- general LP64 related changes
- don't force cast argument to __builtin_arc_sr()
- irqflags: STAT32 is 32-bits but generic raw_local_irq_* expect flags to be long

+++ arch/arc/include/asm/mmu.h
@@ -11,7 +11,7 @@
 #include <linux/threads.h>	/* NR_CPUS */
 
 typedef struct {
-	unsigned long asid[NR_CPUS];	/* 8 bit MMU PID + Generation cycle */
+	unsigned int asid[NR_CPUS];	/* 8 bit MMU PID + Generation cycle */
 } mm_context_t;
 
 #endif
